+++ package.json
@@ -67,10 +67,11 @@
     ]
   },
   "devDependencies": {
-    "@babel/core": "^7.0.0-beta.46",
-    "@babel/preset-env": "^7.0.0-beta.46",
-    "@babel/preset-stage-2": "^7.0.0-beta.46",
-    "babel-jest": "^22.4.3",
+    "@babel/core": "^7.0.0-beta.52",
+    "@babel/preset-env": "^7.0.0-beta.52",
+    "@babel/preset-stage-2": "^7.0.0-beta.52",
+    "babel-core": "^7.0.0-bridge.0",
+    "babel-jest": "^22.4.4",
     "cross-env": "^5.1.4",
     "eslint": "^4.19.1",
     "gitbook-cli": "^2.3.2",
@@ -79,7 +80,7 @@
     "regenerator-runtime": "^0.11.1",
     "rimraf": "^2.6.2",
     "rollup": "^0.58.2",
-    "rollup-plugin-babel": "^4.0.0-beta.4",
+    "rollup-plugin-babel": "^4.0.0-beta.7",
     "rollup-plugin-node-resolve": "^3.3.0",
     "rollup-plugin-replace": "^2.0.0",
     "rollup-plugin-uglify": "^3.0.0",
