+++ src/utils/actor.js
@@ -1,3 +1,6 @@
+import { assertIsValidActor } from './errors'
+
+
 const sharedToStringMethod = function() {
   return this.type
 }
@@ -56,4 +59,28 @@ export function createActorCreator(createActorImplementation) {
 
     return createActorImplementation(actionType)
   }
+}
+
+
+/**
+  Pulls the string action type out of an actor or returns
+  a given string action type as-is.
+*/
+export function extractActionType(actor, method) {
+
+  // The "actor" may be a literal action type string
+  if (typeof actor === 'string') return actor
+
+  assertIsValidActor(actor, method)
+
+  return actor.type
+}
+
+
+/**
+  Pulls the string action types out of a list of (possibly) mixed
+  actors and string action types.
+*/
+export function extractActionTypes(actors, method) {
+  return actors.map(extractActionType, method)
 }
